Move parallel_commands file to util directory
